refactor: move MetricsServerError to domain.exceptions.critical
- Move MetricsServerError from infrastructure to domain.exceptions.critical
- Remove infrastructure/observability/exceptions.py (exception now in domain)
- Update all imports across layers to use domain.exceptions
- Add MetricsServerError to domain.__all__ and domain.exceptions.__all__

This change properly respects layer boundaries:
- Domain layer owns the exception (value object, importable by all layers)
- Infrastructure, application, composition, and interfaces all import from domain
- Eliminates the architectural violation where application imported from infrastructure

MetricsServerError is now a CriticalError subclass with error_type=DB_UNAVAILABLE,
consistent with other infrastructure startup errors.
