+++ src/bioetl/domain/exceptions/critical.py
@@ -147,3 +147,36 @@ class InvalidStateError(CriticalError):
         self.current_state = current_state
         self.attempted_operation = attempted_operation
         super().__init__(message)
+
+
+class MetricsServerError(CriticalError):
+    """Raised when metrics server fails to start with fail_fast=True.
+
+    This is a CRITICAL error - if fail_fast is enabled, the pipeline
+    should not start without operational metrics collection.
+
+    This exception is raised by the Prometheus metrics server when it cannot
+    bind to the specified port and fail_fast mode is enabled.
+
+    Attributes:
+        port: Port that failed to bind.
+        reason: Reason for failure (e.g., "port_in_use", "os_error", "unexpected").
+        original_error: Underlying exception that caused the failure.
+    """
+
+    error_type = ErrorType.DB_UNAVAILABLE
+
+    def __init__(
+        self, port: int, reason: str, original_error: Exception | None = None
+    ) -> None:
+        """Initialize MetricsServerError.
+
+        Args:
+            port: Port that failed.
+            reason: Reason for failure.
+            original_error: Underlying exception.
+        """
+        self.port = port
+        self.reason = reason
+        self.original_error = original_error
+        super().__init__(f"Failed to start metrics server on port {port}: {reason}")
