+++ src/bioetl/infrastructure/observability/__init__.py
@@ -14,7 +14,7 @@ with mandatory fields: ts, level, run_id, pipeline, stage.
 
 from __future__ import annotations
 
-from bioetl.infrastructure.observability.exceptions import MetricsServerError
+from bioetl.domain.exceptions import MetricsServerError
 from bioetl.infrastructure.observability.logging import StructlogLogger
 from bioetl.infrastructure.observability.metrics_server_adapter import (
     MetricsServerAdapter,
