+++ src/bioetl/interfaces/observability.py
@@ -1,21 +1,17 @@
 """Observability interface for BioETL.
 
-Re-exports observability components from the infrastructure layer.
-This module provides a clean interface for external consumers.
+Re-exports observability components for external consumers.
 
 Note:
-    MetricsServerError and start_metrics_server are now imported from
-    infrastructure.observability (not composition._bootstrap) to respect
-    layer boundaries. Interfaces may import from infrastructure since
-    exceptions are value objects.
+    MetricsServerError is defined in domain.exceptions (value object,
+    can be imported by all layers). start_metrics_server is re-exported
+    from infrastructure.observability.
 """
 
 from __future__ import annotations
 
-from bioetl.infrastructure.observability import (
-    MetricsServerError,
-    start_metrics_server,
-)
+from bioetl.domain.exceptions import MetricsServerError
+from bioetl.infrastructure.observability import start_metrics_server
 
 __all__ = [
     "MetricsServerError",
