+++ tests/unit/interfaces/test_observability.py
@@ -5,13 +5,14 @@ from unittest import mock
 import pytest
 
 from bioetl.composition._bootstrap import observability as composition_observability
+from bioetl.domain import exceptions as domain_exceptions
 from bioetl.infrastructure import observability as infra_observability
 from bioetl.infrastructure.observability import server as obs_server
 from bioetl.interfaces import observability
 
 # This module tests the observability interface.
-# The observability module re-exports metrics server components from infrastructure layer
-# (not from composition._bootstrap) to respect layer boundaries.
+# MetricsServerError is now defined in domain.exceptions and re-exported by all layers.
+# start_metrics_server is defined in infrastructure.observability.server.
 
 
 # Mock `_SERVER_STARTED` to isolate state among test cases.
@@ -50,11 +51,13 @@ def test_start_metrics_server_failure():
 def test_interface_re_exports_from_infrastructure():
     """Verify interface re-exports start_metrics_server from infrastructure layer.
 
-    interfaces/observability.py imports from infrastructure.observability
-    (not composition._bootstrap) to respect layer boundaries.
+    interfaces/observability.py imports start_metrics_server from
+    infrastructure.observability to provide the function to external consumers.
     """
     # The function should be the same object since it's a re-export
-    assert observability.start_metrics_server is infra_observability.start_metrics_server
+    assert (
+        observability.start_metrics_server is infra_observability.start_metrics_server
+    )
     # Composition layer also imports from infrastructure, so all should be the same
     assert (
         observability.start_metrics_server
@@ -63,11 +66,18 @@ def test_interface_re_exports_from_infrastructure():
 
 
 def test_interface_exposes_metrics_server_error():
-    """Verify MetricsServerError is exported from interface via infrastructure."""
-    # Interface imports directly from infrastructure
-    assert observability.MetricsServerError is infra_observability.MetricsServerError
-    # All layers should reference the same exception class from infrastructure
-    assert observability.MetricsServerError is obs_server.MetricsServerError
+    """Verify MetricsServerError is exported from interface via domain.
+
+    MetricsServerError is defined in domain.exceptions.critical and
+    re-exported by all layers (infrastructure, composition, interfaces).
+    """
+    # All layers should reference the same exception class from domain
+    assert observability.MetricsServerError is domain_exceptions.MetricsServerError
     assert (
-        observability.MetricsServerError is composition_observability.MetricsServerError
+        infra_observability.MetricsServerError is domain_exceptions.MetricsServerError
+    )
+    assert obs_server.MetricsServerError is domain_exceptions.MetricsServerError
+    assert (
+        composition_observability.MetricsServerError
+        is domain_exceptions.MetricsServerError
     )
